requirements.txt
streamlit==0.84.1
numpy==1.20.1
pandas==1.2.4
matplotlib==3.4.1
altair==4.1.0
imblearn==0.0
xgboost==1.4.0
scikit-learn==0.24.1
scipy==1.6.2
streamlit_wordcloud
nltk
PySastrawi